add basic pytest

+++ Makefile
@@ -15,10 +15,6 @@ BROWSER := python -c "$$BROWSER_PYSCRIPT"
 PULUMI_SKIP_UPDATE_CHECK=true
 PULUMI_DIY_BACKEND_GZIP=true
 
-# use python implementation of protoc to workaround different protoc versions in python-pulumi
-# see https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates
-# PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python
-
 .PHONY: help
 help:
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | \
@@ -63,7 +59,19 @@ py-clean: ## Remove python related artifacts
 .PHONY: buildenv-clean
 buildenv-clean: py-clean ## Remove build environment artifacts
 	@echo "+++ $@"
-	rm -rf .venv
+	if test -e .venv; then rm -rf .venv; fi
+	if test -e infra_shared.egg-info; then rm -rf infra_shared.egg-info; fi
+
+.PHONY: pytest
+pytest: buildenv ## Run pytest Tests
+	@echo "+++ $@"
+	mkdir -p build/tests/
+	. .venv/bin/activate && pytest tests/
+
+.PHONY: pytest-clean
+pytest-clean: ## Remove pytest Artifacts
+	@echo "+++ $@"
+	if test -d $(ROOTDIR)/build/tests; then rm -rf $(ROOTDIR)/build/tests; fi
 
 .PHONY: test-scripts
 test-scripts: buildenv ## Run script Tests
@@ -73,56 +81,55 @@ test-scripts: buildenv ## Run script Tests
 .PHONY: test-sim
 test-sim: buildenv ## Run sim up Tests
 	@echo "+++ $@"
-	mkdir -p $(ROOTDIR)/build/pulumi $(ROOTDIR)/build/tests
-	git init $(ROOTDIR)/build/tests
+	mkdir -p $(ROOTDIR)/build/test_sim
+	git init $(ROOTDIR)/build/test_sim
 	./scripts/create_skeleton.sh \
-		--project-dir $(ROOTDIR)/build/tests --name-library infra --yes
-
+		--project-dir $(ROOTDIR)/build/test_sim --name-library infra --yes
+	mkdir -p $(ROOTDIR)/build/test_sim/state/pulumi $(ROOTDIR)/build/test_sim/target
 	for i in infra uv.lock .venv; do \
-	    f=$(ROOTDIR)/build/tests/$$i && if test ! -e $$f; then ln -s "../../" $$f; fi \
+	    f=$(ROOTDIR)/build/test_sim/$$i && if test ! -e $$f; then ln -s "../../" $$f; fi \
 	done
-	sed -i -r "s#virtualenv: .venv#virtualenv: ../../.venv#g" $(ROOTDIR)/build/tests/Pulumi.yaml
-	mkdir -p $(ROOTDIR)/build/tests/target && \
-	    cp -r $(ROOTDIR)/examples/safe $(ROOTDIR)/build/tests/target
-	cat $(ROOTDIR)/build/tests/config-template.yaml >> $(ROOTDIR)/build/tests/Pulumi.sim.yaml
-	printf "  tests:safe_showcase_unittest: true\n\n" >> $(ROOTDIR)/build/tests/Pulumi.sim.yaml
-	echo "import target.safe" >> $(ROOTDIR)/build/tests/__main__.py
+	sed -i -r "s#virtualenv: .venv#virtualenv: ../../.venv#g" $(ROOTDIR)/build/test_sim/Pulumi.yaml
+	cp -r $(ROOTDIR)/examples/safe $(ROOTDIR)/build/test_sim/target
+	cat $(ROOTDIR)/build/test_sim/config-template.yaml >> $(ROOTDIR)/build/test_sim/Pulumi.sim.yaml
+	printf "  test_sim:safe_showcase_unittest: true\n\n" >> $(ROOTDIR)/build/test_sim/Pulumi.sim.yaml
+	echo "import target.safe" >> $(ROOTDIR)/build/test_sim/__main__.py
 
-	cd $(ROOTDIR)/build/tests && $(PULUMI) login file://$(ROOTDIR)/build/pulumi
-	cd $(ROOTDIR)/build/tests && \
+	cd $(ROOTDIR)/build/test_sim && $(PULUMI) login file://$(ROOTDIR)/build/test_sim/state/pulumi
+	cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) stack init sim --secrets-provider passphrase
-	cd $(ROOTDIR)/build/tests && \
+	cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) stack "select" "sim"
-	cd $(ROOTDIR)/build/tests && \
+	cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) up --stack "sim" --suppress-outputs --yes
-	cd $(ROOTDIR)/build/tests && \
+	cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) stack output --json --stack "sim" --show-secrets safe_butane | jq -r .this_env
 
 .PHONY: sim__
 sim__: ## Run "pulumi $(args)"
 	@echo "+++ $@"
-	cd $(ROOTDIR)/build/tests &&	PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI_INTERACTIVE) $(args)
+	cd $(ROOTDIR)/build/test_sim &&	PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI_INTERACTIVE) $(args)
 
 .PHONY: test-sim-clean
 test-sim-clean: ## Remove Application Artifacts
 	@echo "+++ $@"
-	if test -d $(ROOTDIR)/build/tests; then \
-	    cd $(ROOTDIR)/build/tests && \
+	if test -d $(ROOTDIR)/build/test_sim; then \
+	    cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) --stack "sim" state unprotect --yes 'urn:pulumi:sim::athome::pkg:authority:CACertFactoryVault$$command:local:Command::ca_factory_vault_ca' || true; \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) --stack "sim" state unprotect --yes 'urn:pulumi:sim::athome::pkg:authority:CACertFactoryVault$$command:local:Command::fake_ca_factory_vault_ca' || true; \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) --stack "sim" state unprotect --yes 'urn:pulumi:sim::athome::pkg:authority:CACertFactoryPulumi$$tls:index/selfSignedCert:SelfSignedCert::ca_factory_root_cert' || true; \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) --stack "sim" state unprotect --yes 'urn:pulumi:sim::athome::pkg:authority:CACertFactoryPulumi$$tls:index/privateKey:PrivateKey::ca_factory_root_key' || true; \
 	fi
-	if test -d $(ROOTDIR)/build/tests; then \
-		cd $(ROOTDIR)/build/tests && \
+	if test -d $(ROOTDIR)/build/test_sim; then \
+		cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" PULUMI_CONTINUE_ON_ERROR=true \
 			$(PULUMI) destroy --stack "sim" --yes || true; \
 	fi
-	if test -d $(ROOTDIR)/build/tests; then \
-		cd $(ROOTDIR)/build/tests && \
+	if test -d $(ROOTDIR)/build/test_sim; then \
+		cd $(ROOTDIR)/build/test_sim && \
 		PULUMI_CONFIG_PASSPHRASE="sim" $(PULUMI) stack rm "sim" --force --yes  || true; \
 	fi
-	rm -rf $(ROOTDIR)/build/tests $(ROOTDIR)/build/pulumi || true
+	rm -rf $(ROOTDIR)/build/test_sim $(ROOTDIR)/build/pulumi || true
 
 .PHONY: docs
 docs: buildenv ## Build docs for local usage
@@ -137,7 +144,7 @@ docs-online-build: buildenv ## Build docs for http serve
 	mkdir -p build/docs-online
 	. .venv/bin/activate && mkdocs build -d build/docs-online -f mkdocs.yml
 	@echo "Finished. serve with"
-	@echo ". .venv/bin/activate && python -m http.server --directory build/state/docs-online"
+	@echo ". .venv/bin/activate && python -m http.server --directory build/docs-online"
 
 .PHONY: docs-serve
 docs-serve: buildenv ## Rebuild and serve docs with autoreload
@@ -150,12 +157,12 @@ docs-clean: ## Remove all generated docs
 	rm -rf build/docs-local build/docs-online
 
 .PHONY: clean
-clean: docs-clean test-sim-clean buildenv-clean  ## Remove all artifacts
+clean: pytest-clean test-sim-clean docs-clean buildenv-clean  ## Remove all artifacts
 	@echo "+++ $@"
 	rm -rf build
 
 .PHONY: test-all-local
-test-all-local: clean provision-local buildenv docs-online-build test-scripts test-sim ## Run all tests using local build deps
+test-all-local: clean provision-local buildenv docs-online-build test-scripts test-sim pytest ## Run all tests using local build deps
 	@echo "+++ $@"
 
 .PHONY: test-all-container
